+++ r6502lib/src/emulator/total_cycles.rs
@@ -1,1 +1,1 @@
-pub type TotalCycles = u32;
+pub type TotalCycles = u64;
