+++ r6502lib/src/run_options.rs
@@ -7,6 +7,6 @@ pub struct RunOptions {
     pub trace: bool,
     pub cycles: bool,
     pub reset: bool,
-    pub stop_after: Option<u32>,
+    pub stop_after: Option<u64>,
     pub machine: Option<String>,
 }
